requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7